rename utils package
