--- app/src/main/java/com/github/odaridavid/dimba/commons/ExceptionHandler.kt
@@ -1,30 +1,0 @@
-package com.github.odaridavid.dimba.commons
-
-import android.content.Context
-import com.github.odaridavid.dimba.R
-import java.net.ConnectException
-import java.net.UnknownHostException
-
-/**
- *
- * Copyright 2020 David Odari
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *            http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- *
- **/
-class ExceptionHandler(val context: Context) {
-
-    fun parse(exception: Exception): String {
-        return when (exception) {
-            is ConnectException -> context.getString(R.string.no_internet_connection)
-            is UnknownHostException -> context.getString(R.string.no_internet_connection)
-            else -> exception.message!!
-        }
-    }
-}
