+++ app/src/main/java/com/github/odaridavid/dimba/repositories/FixturesRepository.kt
@@ -1,6 +1,6 @@
 package com.github.odaridavid.dimba.repositories
 
-import com.github.odaridavid.dimba.commons.ResultState
+import com.github.odaridavid.dimba.utils.ResultState
 import com.github.odaridavid.dimba.models.fixtures.LiveFixture
