+++ app/src/main/java/com/github/odaridavid/dimba/repositories/FixturesRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package com.github.odaridavid.dimba.repositories
 
-import com.github.odaridavid.dimba.commons.ResultState
-import com.github.odaridavid.dimba.commons.Success
-import com.github.odaridavid.dimba.commons.executeNonBlocking
+import com.github.odaridavid.dimba.utils.ResultState
+import com.github.odaridavid.dimba.utils.Success
+import com.github.odaridavid.dimba.utils.executeNonBlocking
 import com.github.odaridavid.dimba.mappers.toEntity
 import com.github.odaridavid.dimba.models.fixtures.LiveFixture
 import com.github.odaridavid.dimba.network.FootballApiService
@@ -24,7 +24,6 @@ class FixturesRepositoryImpl(
     private val api: FootballApiService
 ) : FixturesRepository {
 
-    //TODO Check Live Fixture Status Periodically
     override suspend fun getLiveFixtures(): ResultState<List<LiveFixture>> {
         return executeNonBlocking {
             val response = api.getFixturesInPlay()
