+++ app/src/main/java/com/github/odaridavid/dimba/ui/MainActivity.kt
@@ -1,6 +1,5 @@
 package com.github.odaridavid.dimba.ui
 
-import android.content.SharedPreferences
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
@@ -9,8 +8,6 @@ import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.navigateUp
 import androidx.navigation.ui.setupActionBarWithNavController
 import com.github.odaridavid.dimba.R
-import com.github.odaridavid.dimba.commons.ThemeUtils
-import org.koin.android.ext.android.inject
 
 /**
  *
