+++ app/src/main/java/com/github/odaridavid/dimba/utils/NetworkUtils.kt
@@ -1,4 +1,4 @@
-package com.github.odaridavid.dimba.commons
+package com.github.odaridavid.dimba.utils
 
 import android.content.Context
 import android.net.ConnectivityManager
